AICropRecommendation-2/Agri Doctor/requirements.txt
# Core ML dependencies
torch
torchvision
numpy

# Database
mysql-connector-python>=8.0.0
python-dotenv>=1.0.0

# Image processing
Pillow>=9.0.0
opencv-python>=4.8.0

# Grad-CAM visualization
grad-cam>=1.4.8

# Web framework and API
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Data analysis and visualization
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
pandas>=1.5.0

# Utilities
requests>=2.31.0
pathlib2>=2.3.0

# JSON handling
pydantic>=2.0.0

# Additional support
tqdm>=4.65.0
